+++ map.cub
@@ -1,8 +1,8 @@
+
 NO pics/eagle.png
 SO pics/greystone.png
 WE pics/pngtree.png
 EA pics/wood.png
-
 
 F 34,41,41
 C 7,8,8
@@ -21,5 +21,5 @@ C 7,8,8
 10000000000000001000000010001
 10000000000000000001010010001
 11000001110101011111011110E0111
-11110111 1110101 100000000001 
+11110111 1110101 100000000001
 11111111 1111111 111111111111
