+++ parser/get_identifier.c
@@ -3,61 +3,23 @@
 /*                                                        :::      ::::::::   */
 /*   get_identifier.c                                   :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
-/*   By: mtravez <mtravez@student.42.fr>            +#+  +:+       +#+        */
+/*   By: cmeng <cmeng@student.42.fr>                +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/09/22 11:46:53 by christianme       #+#    #+#             */
-/*   Updated: 2023/09/22 15:09:14 by mtravez          ###   ########.fr       */
+/*   Updated: 2023/10/16 17:27:37 by cmeng            ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "../cub3d.h"
 
-static int	rgb_to_uint(char **frag, uint32_t *color)
-{
-	int	red;
-	int	green;
-	int	blue;
-
-	red = atoi(frag[0]);
-	green = atoi(frag[1]);
-	blue = atoi(frag[2]);
-	if (red < 0 || green < 0 || blue < 0)
-		return (1);
-	if (red > 255 || green > 255 || blue > 255)
-		return (1);
-	*color = ((red << 24) | (green << 16) | (blue << 8) | 255);
-	return (0);
-}
-
-static int	set_color(char *str, uint32_t *color)
-{
-	char	**frag;
-	int		i;
-
-	frag = ft_split(str, ',');
-	i = 0;
-	while (frag[i])
-		i++;
-	if (i != 3)
-		return (free_2d(frag, i), free(str), 1);
-	if (rgb_to_uint(frag, color))
-		return (free_2d(frag, i), free(str), 1);
-	free_2d(frag, i);
-	free(str);
-	return (0);
-}
-
 static int	set_id(t_data *data, char **frag)
 {
-	if (str_eq(frag[0], "EA\0") && data->ea == NULL)
-		data->ea = ft_substr(frag[1], 0, ft_strlen(frag[1]) - 1);
-	else if (str_eq(frag[0], "SO\0") && data->so == NULL)
-		data->so = ft_substr(frag[1], 0, ft_strlen(frag[1]) - 1);
-	else if (str_eq(frag[0], "WE\0") && data->we == NULL)
-		data->we = ft_substr(frag[1], 0, ft_strlen(frag[1]) - 1);
-	else if (str_eq(frag[0], "NO\0") && data->no == NULL)
-		data->no = ft_substr(frag[1], 0, ft_strlen(frag[1]) - 1);
-	else if (str_eq(frag[0], "F\0"))
+	int	i;
+
+	i = 0;
+	while (frag[i] != NULL)
+		i++;
+	if (str_eq(frag[0], "F\0"))
 	{
 		if (set_color(ft_strdup(frag[1]), &data->color_f))
 			return (1);
@@ -67,8 +29,15 @@ static int	set_id(t_data *data, char **frag)
 		if (set_color(ft_strdup(frag[1]), &data->color_c))
 			return (1);
 	}
+	else if ((str_eq(frag[0], "EA\0")) || (str_eq(frag[0], "SO\0"))
+		|| (str_eq(frag[0], "WE\0")) || (str_eq(frag[0], "NO\0")))
+	{
+		if (set_textures(data, frag))
+			return (1);
+	}
 	else
-		return (free(frag), 1);
+		return (1);
+		// return (free_2d(frag, i), 1);
 	return (0);
 }
